refactor(orgtrack): modularize history commands
