+++ src-tauri/src/orgtrack/history_commands/provider_wrappers.rs
@@ -1,0 +1,396 @@
+use super::*;
+
+use super::projection::open_cache_conn;
+
+#[tauri::command]
+pub async fn opencode_history_chunks(
+    session_id: String,
+) -> Result<Vec<core_types::activity::ActivityChunk>, String> {
+    tokio::task::spawn_blocking(move || {
+        opencode_history::load_opencode_history_for_session(&session_id)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn opencode_recent_paths(
+    limit: Option<usize>,
+) -> Result<Vec<opencode_history::OpenCodeRecentPath>, String> {
+    let limit = limit.unwrap_or(20);
+    tokio::task::spawn_blocking(move || {
+        let mut conn = open_cache_conn()?;
+        opencode_history::list_opencode_recent_paths(&mut conn, limit)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn warp_history_chunks(
+    session_id: String,
+) -> Result<Vec<core_types::activity::ActivityChunk>, String> {
+    tokio::task::spawn_blocking(move || warp_history::load_warp_history_for_session(&session_id))
+        .await
+        .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn warp_recent_paths(
+    limit: Option<usize>,
+) -> Result<Vec<warp_history::WarpRecentPath>, String> {
+    let limit = limit.unwrap_or(20);
+    tokio::task::spawn_blocking(move || {
+        let mut conn = open_cache_conn()?;
+        warp_history::list_warp_recent_paths(&mut conn, limit)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn zcode_history_chunks(
+    session_id: String,
+) -> Result<Vec<core_types::activity::ActivityChunk>, String> {
+    tokio::task::spawn_blocking(move || zcode_history::load_zcode_history_for_session(&session_id))
+        .await
+        .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn zcode_recent_paths(
+    limit: Option<usize>,
+) -> Result<Vec<zcode_history::ZCodeRecentPath>, String> {
+    let limit = limit.unwrap_or(20);
+    tokio::task::spawn_blocking(move || {
+        let mut conn = open_cache_conn()?;
+        zcode_history::list_zcode_recent_paths(&mut conn, limit)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn qoder_history_chunks(
+    session_id: String,
+) -> Result<Vec<core_types::activity::ActivityChunk>, String> {
+    tokio::task::spawn_blocking(move || {
+        let conn = open_cache_conn()?;
+        qoder_history::load_qoder_history_for_session(&conn, &session_id)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn qoder_recent_paths(
+    limit: Option<usize>,
+) -> Result<Vec<qoder_history::QoderRecentPath>, String> {
+    let limit = limit.unwrap_or(20);
+    tokio::task::spawn_blocking(move || {
+        let mut conn = open_cache_conn()?;
+        qoder_history::list_qoder_recent_paths(&mut conn, limit)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn mimo_code_history_chunks(
+    session_id: String,
+) -> Result<Vec<core_types::activity::ActivityChunk>, String> {
+    tokio::task::spawn_blocking(move || {
+        let conn = open_cache_conn()?;
+        mimo_code_history::load_mimo_code_history_for_session(&conn, &session_id)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn mimo_code_recent_paths(
+    limit: Option<usize>,
+) -> Result<Vec<mimo_code_history::MimoCodeRecentPath>, String> {
+    let limit = limit.unwrap_or(20);
+    tokio::task::spawn_blocking(move || {
+        let mut conn = open_cache_conn()?;
+        mimo_code_history::list_mimo_code_recent_paths(&mut conn, limit)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn omp_history_chunks(
+    session_id: String,
+) -> Result<Vec<core_types::activity::ActivityChunk>, String> {
+    tokio::task::spawn_blocking(move || {
+        let conn = open_cache_conn()?;
+        omp_history::load_omp_history_for_session(&conn, &session_id)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn omp_recent_paths(
+    limit: Option<usize>,
+) -> Result<Vec<omp_history::OmpRecentPath>, String> {
+    let limit = limit.unwrap_or(20);
+    tokio::task::spawn_blocking(move || {
+        let mut conn = open_cache_conn()?;
+        omp_history::list_omp_recent_paths(&mut conn, limit)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn pi_history_chunks(
+    session_id: String,
+) -> Result<Vec<core_types::activity::ActivityChunk>, String> {
+    tokio::task::spawn_blocking(move || {
+        let conn = open_cache_conn()?;
+        pi_history::load_pi_history_for_session(&conn, &session_id)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn pi_recent_paths(
+    limit: Option<usize>,
+) -> Result<Vec<pi_history::PiRecentPath>, String> {
+    let limit = limit.unwrap_or(20);
+    tokio::task::spawn_blocking(move || {
+        let mut conn = open_cache_conn()?;
+        pi_history::list_pi_recent_paths(&mut conn, limit)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn qoder_cli_history_chunks(
+    session_id: String,
+) -> Result<Vec<core_types::activity::ActivityChunk>, String> {
+    tokio::task::spawn_blocking(move || {
+        let conn = open_cache_conn()?;
+        qoder_cli_history::load_qoder_cli_history_for_session(&conn, &session_id)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn qoder_cli_recent_paths(
+    limit: Option<usize>,
+) -> Result<Vec<qoder_cli_history::QoderCliRecentPath>, String> {
+    let limit = limit.unwrap_or(20);
+    tokio::task::spawn_blocking(move || {
+        let mut conn = open_cache_conn()?;
+        qoder_cli_history::list_qoder_cli_recent_paths(&mut conn, limit)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn qwen_code_history_chunks(
+    session_id: String,
+) -> Result<Vec<core_types::activity::ActivityChunk>, String> {
+    tokio::task::spawn_blocking(move || {
+        let conn = open_cache_conn()?;
+        qwen_code_history::load_qwen_code_history_for_session(&conn, &session_id)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn qwen_code_recent_paths(
+    limit: Option<usize>,
+) -> Result<Vec<qwen_code_history::QwenCodeRecentPath>, String> {
+    let limit = limit.unwrap_or(20);
+    tokio::task::spawn_blocking(move || {
+        let mut conn = open_cache_conn()?;
+        qwen_code_history::list_qwen_code_recent_paths(&mut conn, limit)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn kimi_history_chunks(
+    session_id: String,
+) -> Result<Vec<core_types::activity::ActivityChunk>, String> {
+    tokio::task::spawn_blocking(move || {
+        let conn = open_cache_conn()?;
+        kimi_history::load_kimi_history_for_session(&conn, &session_id)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn kimi_recent_paths(
+    limit: Option<usize>,
+) -> Result<Vec<kimi_history::KimiRecentPath>, String> {
+    let limit = limit.unwrap_or(20);
+    tokio::task::spawn_blocking(move || {
+        let mut conn = open_cache_conn()?;
+        kimi_history::list_kimi_recent_paths(&mut conn, limit)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn windsurf_history_chunks(
+    session_id: String,
+) -> Result<Vec<core_types::activity::ActivityChunk>, String> {
+    tokio::task::spawn_blocking(move || {
+        windsurf_history::load_windsurf_history_for_session(&session_id)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn windsurf_recent_paths(
+    limit: Option<usize>,
+) -> Result<Vec<windsurf_history::WindsurfRecentPath>, String> {
+    let limit = limit.unwrap_or(20);
+    tokio::task::spawn_blocking(move || {
+        let mut conn = open_cache_conn()?;
+        windsurf_history::list_windsurf_recent_paths(&mut conn, limit)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn trae_history_chunks(
+    session_id: String,
+) -> Result<Vec<core_types::activity::ActivityChunk>, String> {
+    tokio::task::spawn_blocking(move || {
+        let conn = open_cache_conn()?;
+        trae_history::load_trae_history_for_session(&conn, &session_id)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn trae_recent_paths(
+    limit: Option<usize>,
+) -> Result<Vec<trae_history::TraeRecentPath>, String> {
+    let limit = limit.unwrap_or(20);
+    tokio::task::spawn_blocking(move || {
+        let mut conn = open_cache_conn()?;
+        trae_history::list_trae_recent_paths(&mut conn, limit)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn cline_history_chunks(
+    session_id: String,
+) -> Result<Vec<core_types::activity::ActivityChunk>, String> {
+    tokio::task::spawn_blocking(move || {
+        let conn = open_cache_conn()?;
+        cline_history::load_cline_history_for_session(&conn, &session_id)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn cline_recent_paths(
+    limit: Option<usize>,
+) -> Result<Vec<cline_history::ClineRecentPath>, String> {
+    let limit = limit.unwrap_or(20);
+    tokio::task::spawn_blocking(move || {
+        let mut conn = open_cache_conn()?;
+        cline_history::list_cline_recent_paths(&mut conn, limit)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn workbuddy_history_chunks(
+    session_id: String,
+) -> Result<Vec<core_types::activity::ActivityChunk>, String> {
+    tokio::task::spawn_blocking(move || {
+        let conn = open_cache_conn()?;
+        workbuddy_history::load_workbuddy_history_for_session(&conn, &session_id)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[tauri::command]
+pub async fn workbuddy_recent_paths(
+    limit: Option<usize>,
+) -> Result<Vec<workbuddy_history::WorkBuddyRecentPath>, String> {
+    let limit = limit.unwrap_or(20);
+    tokio::task::spawn_blocking(move || {
+        let mut conn = open_cache_conn()?;
+        workbuddy_history::list_workbuddy_recent_paths(&mut conn, limit)
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
+
+#[derive(serde::Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ExternalHistorySourceStatsWire {
+    pub source_id: String,
+    pub session_count: usize,
+    pub subagent_count: usize,
+    pub last_used_at: Option<String>,
+}
+
+/// One compact cache-only inventory read for every requested source. This
+/// command never opens provider databases or walks transcript directories.
+#[tauri::command]
+pub async fn external_history_source_stats(
+    sources: Vec<String>,
+) -> Result<Vec<ExternalHistorySourceStatsWire>, String> {
+    let mut seen_sources = HashSet::with_capacity(sources.len());
+    for source in &sources {
+        if !seen_sources.insert(source.clone()) {
+            return Err(format!("Duplicate external history source: {source}"));
+        }
+        if !imported_history::metadata::is_imported_history_source(source) {
+            return Err(format!("Unknown external history source: {source}"));
+        }
+    }
+
+    tokio::task::spawn_blocking(move || {
+        let conn = open_cache_conn()?;
+        let cached = imported_history::cache::all_source_stats_from_conn(&conn)?
+            .into_iter()
+            .map(|stats| (stats.source.clone(), stats))
+            .collect::<std::collections::HashMap<_, _>>();
+        Ok(sources
+            .into_iter()
+            .map(|source_id| {
+                let stats = cached.get(&source_id);
+                ExternalHistorySourceStatsWire {
+                    source_id,
+                    session_count: stats.map_or(0, |row| row.session_count),
+                    subagent_count: stats.map_or(0, |row| row.subagent_count),
+                    last_used_at: stats
+                        .and_then(|row| row.last_used_at_ms)
+                        .and_then(chrono::DateTime::<chrono::Utc>::from_timestamp_millis)
+                        .map(|value| value.to_rfc3339()),
+                }
+            })
+            .collect())
+    })
+    .await
+    .map_err(|err| format!("Task join error: {err}"))?
+}
